+++ src/Actions/MakeSchemaForReflectionTypeTest.php
@@ -85,13 +85,13 @@ it('throws an exception if the type is not supported', function () {
     $action = new MakeSchemaForReflectionType;
 
     $action->handle($wrapper, 'unsupportedProperty');
-})->throws(\Exception::class, 'JSON Schema transformation is not supported for the "Closure" type.');
+})->throws(Exception::class, 'JSON Schema transformation is not supported for the "Closure" type.');
 
 it('throws an exception if the enum is not backed', function () {
     $wrapper = PropertyWrapper::make(MakeSchemaForReflectionTypeTest::class, 'nonBackedEnumProperty')->getReflectionType();
     $action = new MakeSchemaForReflectionType;
 
-    expect(fn () => $action->handle($wrapper, 'nonBackedEnumProperty'))->toThrow(\Exception::class, 'Enum "NonBackedEnum" is not a backed enum.');
+    expect(fn () => $action->handle($wrapper, 'nonBackedEnumProperty'))->toThrow(Exception::class, 'Enum "NonBackedEnum" is not a backed enum.');
 });
 
 it('makes a UnionSchema for a nullable type by default', function () {
@@ -114,4 +114,4 @@ it('throws an exception if an intersection type is passed', function () {
     $action = new MakeSchemaForReflectionType;
 
     $action->handle($wrapper, 'unsupportedType');
-})->throws(\Exception::class, 'JSON Schema transformation is not supported for intersection types.');
+})->throws(Exception::class, 'JSON Schema transformation is not supported for intersection types.');