+++ src/Support/DocBlockParser.php
@@ -2,9 +2,11 @@
 
 namespace BasilLangevin\LaravelDataJsonSchemas\Support;
 
+use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
 use PHPStan\PhpDocParser\Ast\PhpDoc\ParamTagValueNode;
+use PHPStan\PhpDocParser\Ast\PhpDoc\PhpDocChildNode;
 use PHPStan\PhpDocParser\Ast\PhpDoc\PhpDocNode;
 use PHPStan\PhpDocParser\Ast\PhpDoc\PhpDocTextNode;
 use PHPStan\PhpDocParser\Ast\PhpDoc\VarTagValueNode;
@@ -55,7 +57,7 @@ class DocBlockParser
     /**
      * Get the children of the doc block.
      *
-     * @return \PHPStan\PhpDocParser\Ast\PhpDoc\PhpDocChildNode[]
+     * @return PhpDocChildNode[]
      */
     protected function getChildren(): array
     {
@@ -65,7 +67,7 @@ class DocBlockParser
     /**
      * Get the text nodes of the doc block.
      *
-     * @return \PHPStan\PhpDocParser\Ast\PhpDoc\PhpDocTextNode[]
+     * @return PhpDocTextNode[]
      */
     protected function getTextNodes(): array
     {
@@ -73,6 +75,14 @@ class DocBlockParser
             $this->getChildren(),
             fn ($child) => $child instanceof PhpDocTextNode
         );
+    }
+
+    /**
+     * Get the first text node of the doc block.
+     */
+    protected function getFirstTextNode(): ?PhpDocTextNode
+    {
+        return Arr::first($this->getTextNodes());
     }
 
     /**
@@ -84,7 +94,7 @@ class DocBlockParser
             return null;
         }
 
-        $content = str($this->getTextNodes()[0]->text);
+        $content = str($this->getFirstTextNode()?->text ?? '');
 
         if ($content->contains("\n\n")) {
             $content = $content->before("\n\n")->trim();
@@ -116,7 +126,7 @@ class DocBlockParser
 
         $summary = $this->getSummary() ?? ''; // @pest-mutate-ignore Strict type assertion for PHPStan.
 
-        $description = str($this->getTextNodes()[0]->text)
+        $description = str($this->getFirstTextNode()?->text ?? '')
             ->after($summary)
             ->trim()
             ->toString();
@@ -135,7 +145,7 @@ class DocBlockParser
     /**
      * Get a Collection of the doc block's param tags.
      *
-     * @return \Illuminate\Support\Collection<int, \PHPStan\PhpDocParser\Ast\PhpDoc\ParamTagValueNode>
+     * @return Collection<int, ParamTagValueNode>
      */
     protected function getParamTagValues(): Collection
     {
@@ -164,7 +174,7 @@ class DocBlockParser
     /**
      * Get a Collection of the doc block's var tags.
      *
-     * @return \Illuminate\Support\Collection<int, \PHPStan\PhpDocParser\Ast\PhpDoc\VarTagValueNode>
+     * @return Collection<int, VarTagValueNode>
      */
     protected function getVarTagValues(): Collection
     {